requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
twilio==8.10.0
python-dotenv==1.0.0
requests==2.31.0 